Improve commit connection validation
